backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4
python-dotenv==1.0.1
openai==1.58.1
pytest==8.3.4
httpx==0.28.1
PyPDF2==3.0.1
python-multipart==0.0.9
